kvparams extract configuration params
