+++ pkg/kv/cosmosdb/main_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 )
 
 var testParams *kvparams.CosmosDB
