+++ pkg/kv/dynamodb/main_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 	"github.com/treeverse/lakefs/pkg/testutil"
 )
