+++ pkg/kv/dynamodb/store_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/treeverse/lakefs/pkg/kv/dynamodb"
 	"github.com/treeverse/lakefs/pkg/kv/kvtest"
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 )
 
 func TestDynamoKV(t *testing.T) {
