+++ pkg/kv/local/driver.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/dgraph-io/badger/v3"
 	"github.com/treeverse/lakefs/pkg/kv"
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 	"github.com/treeverse/lakefs/pkg/logging"
 )
