+++ pkg/kv/local/store_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/treeverse/lakefs/pkg/kv/kvtest"
 	"github.com/treeverse/lakefs/pkg/kv/local"
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 )
 
 func TestLocalKV(t *testing.T) {
