+++ pkg/kv/mem/store.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 
 	"github.com/treeverse/lakefs/pkg/kv"
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 )
 
 type Driver struct{}
