+++ pkg/kv/migration.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	kvparams "github.com/treeverse/lakefs/pkg/kv/params"
+	"github.com/treeverse/lakefs/pkg/kv/kvparams"
 	"github.com/treeverse/lakefs/pkg/logging"
 )
